pyproject.toml
[project]
name = "youtube-downloader-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "SamuelBFavarin",email = "samuelbfavarin@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp (>=2025.7.21,<2026.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "mcp-server (>=0.1.4,<0.2.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
